+++ tests/test_searcher.py
@@ -5,11 +5,9 @@ from init_board_helper import board_setup, init_board, score_fen
 from sporkfish.board.board_factory import BoardFactory, BoardPyChess
 from sporkfish.evaluator import Evaluator
 from sporkfish.searcher.minimax import MiniMaxVariants
-from sporkfish.searcher.move_ordering.move_order_heuristic import (
-    KillerMoveHeuristic,
-    MoveOrderMode,
-    MvvLvaHeuristic,
-)
+from sporkfish.searcher.move_ordering.killer_move_heuristic import KillerMoveHeuristic
+from sporkfish.searcher.move_ordering.move_order_heuristic import MoveOrderMode
+from sporkfish.searcher.move_ordering.mvv_lva_heuristic import MvvLvaHeuristic
 from sporkfish.searcher.searcher import Searcher
 from sporkfish.searcher.searcher_config import SearcherConfig
 from sporkfish.searcher.searcher_factory import SearcherFactory
